requirements.txt
Flask
python-dotenv
boto3
django
djangorestframework
django-cors-headers
requests
pyjwt
jsonify
cryptography